Initial commit with prototype
Creates a prototype of an Operational Task Manager. The task manager
focuses on a single goal at a time. All goals are prioritised and
failing to meet one means failing to meet the remaining.

The prototype is crude and the following improvements can be made:
1. Nicer interface than the current command line print out
2. Postpone a goal, keeping the requirement that all subsequent goals
fail, if the postponed goal is not resolved
3. Pomodoro timer in order to introduce breaks at appropriate times.
Could be made slacker so that the break only starts after a task has
been completed

Scope: Global